feat:[TRI-281] - some refactor
