+++ irs-api/src/main/java/net/catenax/irs/aaswrapper/submodel/domain/SubmodelTestdataCreator.java
@@ -16,12 +16,10 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import net.catenax.irs.aaswrapper.dto.LifecycleContextCharacteristic;
-
 /**
  * Class to create Submodel Testdata
  */
-public class SubmodelTestdataCreator {
+class SubmodelTestdataCreator {
     private final List<AssemblyPartRelationship> testData;
 
     public SubmodelTestdataCreator() {
@@ -57,8 +55,7 @@ public class SubmodelTestdataCreator {
 
     public AssemblyPartRelationship createDummyAssemblyPartRelationshipForId(final String catenaXId) {
         final List<AssemblyPartRelationship> collect = testData.stream()
-                                                               .filter(assemblyPartRelationship -> assemblyPartRelationship.getCatenaXId()
-                                                                                                                           .equals(catenaXId))
+                                                               .filter(assemblyPartRelationship -> assemblyPartRelationship.getCatenaXId().equals(catenaXId))
                                                                .collect(Collectors.toList());
         final AssemblyPartRelationship other = new AssemblyPartRelationship();
         other.setCatenaXId(catenaXId);
